backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-socketio==5.7.2
python-engineio==4.3.4
websockets==10.4
numpy==1.24.2
scipy==1.10.1
pyrtlsdr==0.2.92
pyusb==1.2.1
pydantic==1.10.7
python-multipart==0.0.6
aiofiles==23.1.0 